refactor(phonemes): share syllable allocation with project model
