+++ src/app/Modules/Inference/Tasks/Syllabification.h
@@ -4,20 +4,12 @@
 #include "Modules/Inference/Models/NoteInferenceSnapshot.h"
 #include "Modules/Inference/Models/PhonemeNameResult.h"
 
-#include <QStringList>
+#include <lite/ProjectModel/Utils/Syllabification.h>
 
 class InferInputNote;
 class Timeline;
 
 namespace Syllabification {
-    struct PhonemeRange {
-        int start = 0;
-        int count = 0;
-    };
-
-    bool isSyllabificationLyric(const QString &lyric);
-    QList<PhonemeRange> phonemeRangesForNotes(const QStringList &lyrics,
-                                              const QList<PhonemeName> &phonemes);
     void keepPhonemesOnWordRoots(const QList<NoteInferenceSnapshot> &notes,
                                  QList<PhonemeNameResult> &results);
     void distributeForInference(const QStringList &lyrics, QList<InferInputNote> &notes,
